% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{chronology}
\alias{chronology}
\title{chronology}
\usage{
chronology(titles.vec, patterns, starts, freqs, chronDir, overlay = FALSE)
}
\arguments{
\item{titles.vec}{character vector containing the titles of the stories}

\item{patterns}{vector of character string query patterns}

\item{starts}{integer vector of starting positions}

\item{freqs}{return value of frequency()}

\item{chronDir}{character string full path name for output directory}

\item{overlay}{Boolean if TRUE overlay the chronolgy for multiple search patterns}
}
\value{
returns no value, but has side effect generating graph
}
\description{
frequencies plotted in order of date
(if the titles are given in order of date)
}
\examples{
freqDir<-tempdir()
chronDir<-sprintf("\%s/chronology",freqDir)
dir.create(chronDir)
dir.create(sprintf("\%s/plots",chronDir))
dir.create(sprintf("\%s/archive",chronDir))
print(chronDir)
chr<-chronology(titles.vec,c("Holmes","Watson"),starts,freqs,chronDir)

}
