% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{concordance}
\alias{concordance}
\title{concordance}
\usage{
concordance(freqs, titles.vec, texts.vec, starts, window, odir)
}
\arguments{
\item{freqs}{return value of frequency()}

\item{titles.vec}{character vector containing the titles of the stories}

\item{texts.vec}{character vector of entire text}

\item{starts}{integer vector of starting positions}

\item{window}{integer number of lines to take before and after the pattern match}

\item{odir}{character string containing the full path name for the output directory}
}
\value{
returns no value but has side effect of generating graphs
}
\description{
retrieve words that are close to occurrences of pattern
}
\examples{
\donttest{
con<-concordance(freqs,titles.vec[3],texts.vec,starts,window=2,odir=tempdir())
}

}
