% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{contingency}
\alias{contingency}
\title{contingency}
\usage{
contingency(inside, outside)
}
\arguments{
\item{inside}{numeric vector of raw counts}

\item{outside}{numeric vector of raw counts}
}
\value{
numeric vector of chisq.test() p.values
}
\description{
compute chisq value for a 2 x 2 contingency table
}
\examples{
con<-contingency(inside=c(4,5),outside=c(20,7))

}
