% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{distributions}
\alias{distributions}
\title{distributions}
\usage{
distributions(freqs, titles.vec, minl, P, odir)
}
\arguments{
\item{freqs}{return value of frequency()}

\item{titles.vec}{character vector containing the titles of the stories}

\item{minl}{is an integer param passed to dpseg::dpseg}

\item{P}{is a numeric param passed to dpseg::dpseg}

\item{odir}{character string containing the full path name for the output directory}
}
\value{
returns no value but has side effect of generating graphs
}
\description{
compute distribution of ratio of number of occurrences of query string divided by total number of words
}
\examples{
dis<-distributions(freqs,titles.vec[1],minl=100,P=0.00001,tempdir())

}
