% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{freqHist}
\alias{freqHist}
\title{freqHist}
\usage{
freqHist(patterns, starts, titles.vec, freqs, histDir)
}
\arguments{
\item{patterns}{vector of character string query patterns}

\item{starts}{integer vector of starting positions}

\item{titles.vec}{character vector containing the titles of the stories}

\item{freqs}{return value of frequency()}

\item{histDir}{character string full path name for output directory}
}
\value{
returns no value, but has side effect generating histogram
}
\description{
histogram of frequencies
}
\examples{
fh<-freqHist(patterns,starts,titles.vec,freqs,histDir=tempdir())

}
