% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{frequency}
\alias{frequency}
\title{frequency}
\usage{
frequency(texts.vec, starts, patterns)
}
\arguments{
\item{texts.vec}{character vector of entire text}

\item{starts}{integer vector of starting positions}

\item{patterns}{vector of character string query patterns}
}
\value{
a list whose components are sub-lists
\itemize{ # indexed by the titles of the stories
\item start integer starting line in text
\item end integer ending line in text
\item wPerLine integer words perline
\item wordSum integer sum of wPerLine
\item patterns a sub-list
\itemize{
\item integer pPerLine integer patterns per line
\item patSum integer total of pPerLine
\item fraction numeric ratio of patSum/wordSum
}
}
}
\description{
compute ratio of number of occurrences of query string divided by total number of words
}
\examples{
fr<-frequency(texts.vec,starts,patterns)
}
