% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{grabFunctionParameters}
\alias{grabFunctionParameters}
\title{grabFunctionParameters}
\usage{
grabFunctionParameters()
}
\value{
a list whose components are the symbolic names of the function parameters,
and their values.
}
\description{
retrieve capture all of the parameter names and values passed in
}
\details{
copied and pasted from
https://stackoverflow.com/questions/66329835/using-r-how-to-get-all-parameters-passed-into-a-function-with-their-values
}
