% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{lengths}
\alias{lengths}
\title{lengths}
\usage{
lengths(titles.vec, patterns, starts, freqs, lengthDir)
}
\arguments{
\item{titles.vec}{character vector containing the titles of the stories}

\item{patterns}{vector of character string query patterns}

\item{starts}{integer vector of starting positions}

\item{freqs}{return value of frequency()}

\item{lengthDir}{character string full path name for output directory}
}
\value{
returns no value, but has side effect generating graph
}
\description{
frequencies plotted in order of story length
}
\examples{
freqDir<-tempdir()
lengthDir<-sprintf("\%s/length",freqDir)
dir.create(lengthDir)
print(lengthDir)
dir.create(sprintf("\%s/plots",lengthDir))
dir.create(sprintf("\%s/archive",lengthDir))
le<-lengths(titles.vec,patterns,starts,freqs,lengthDir) 

}
