% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{mergeTables}
\alias{mergeTables}
\title{mergeTables}
\usage{
mergeTables(tv, tw, cnv, cnw)
}
\arguments{
\item{tv}{first table}

\item{tw}{second table}

\item{cnv}{character name for column coming from v}

\item{cnw}{character name for column coming from w}
}
\value{
numeric matrix generated from merging tables from v and w
}
\description{
merge (inner join) the results in 2 tables generated from 2 vectors
}
\examples{
mt<-mergeTables(inside,outside,"in","out")[1:10,]

}
