% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{plot_dpseg2}
\alias{plot_dpseg2}
\title{plot_dpseg2}
\usage{
plot_dpseg2(
  x,
  delog = FALSE,
  col,
  main,
  xlab,
  ylab,
  res = 10,
  vlines,
  overlay,
  textX,
  textY,
  textLabel,
  ylim
)
}
\arguments{
\item{x}{dpseg object to plot}

\item{delog}{Boolean use log scale if TRUE}

\item{col}{color}

\item{main}{character title of graph}

\item{xlab}{character label for x axis}

\item{ylab}{character label for y axis}

\item{res}{numeric resolution}

\item{vlines}{Boolean if FALSE suppress vertical lines in graph}

\item{overlay}{Boolean if TRUE this plot is an overlay of previous plot}

\item{textX}{numeric x position for text box}

\item{textY}{numeric y position for text box}

\item{textLabel}{character string to label the points in the graph}

\item{ylim}{numeric vector ylim for plot}
}
\value{
returns no value but has side effect of producing a graph
}
\description{
Alternative plot procedure for dpseg,
special function provided personally by dpseg curator.
I made a few custom tweeks
Including option to overlay multiple plots
}
\examples{
pdp<-plot_dpseg2(segs,overlay=FALSE,xlab="xaxis",
  ylab="yaxis",vlines=FALSE,textX=2000,textY=20,
  textLabel="label",ylim=c(0,60))

}
