% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{retrieveLmStats}
\alias{retrieveLmStats}
\title{retrieveLmStats}
\usage{
retrieveLmStats(x, y)
}
\arguments{
\item{x}{is second argument to lm()}

\item{y}{is first argument to lm()}
}
\value{
returns a list containing the return value of lm,
intercept, slope, r.squared, and adj.r.squared
}
\description{
This function retrieves intercept, slope,
r.squared, and adj.r.squared from lm()
}
\examples{
retr<-retrieveLmStats(1:10,runif(10,0,1))

}
