% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{segments}
\alias{segments}
\title{segments}
\usage{
segments(segs)
}
\arguments{
\item{segs}{return value of dpseg::dpseg()}
}
\value{
reformatted matrix suitable for printing
}
\description{
reformat seqs$segments as a legend to insert into segment plot
}
\examples{
seg<-segments(segs)

}
