% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{startLine}
\alias{startLine}
\title{startLine}
\usage{
startLine(titles.vec, texts.vec, toupper)
}
\arguments{
\item{titles.vec}{is a character string containing the full path name for a text file containing the titles of the stories in the same order that they appear in the texts file}

\item{texts.vec}{is a character string containing the full path name for a text file containing the full texts of all of the stories}

\item{toupper}{is a Boolean TRUE if the titles should be converted to upper case}
}
\value{
an integer vector of the starting lines of each story
}
\description{
where does each story start?
}
\details{
each title in titles.vec must appear on a single line
in titles.vec and texts.vec -
a title cannot be split across multiple lines.
each title must only appear one time within titles.vec and texts.vec
}
\examples{
sl<-startLine(titles.vec,texts.vec,toupper=TRUE)
}
