% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sherlock.R
\name{strSplitTab}
\alias{strSplitTab}
\title{strSplitTab}
\usage{
strSplitTab(t)
}
\arguments{
\item{t}{vector of character strings representing lines of the orginal text}
}
\value{
a sorted table of raw word counts
}
\description{
use strsplit to parse words from text t, delete the empty string
from the result, and compile into a sorted table of word frequencies
}
\examples{
sst<-strSplitTab(texts.vec)
 
}
